The Grand Refactor: shrink main()
We moved much of the code out of `main()` and into a new package, `bench`.
`main()`'s primary purpose is to parse flags and call `bench`'s tests.

Also we moved main into the `gobonniego` subdirectory (similar to Onsi's
`ginkgo` in that respect) so that anyone could use _gobonniego_ as a
library.