Reads and Writes are reported additionally in MB
...as well as MiB/s. This is a pre-emptive strike: The distinction
between MiB (2^20) and MB (10^6) is subtle, and if an earnest user does
the calculation, she'll notice that there's a discrepancy, that the MB/s
is 4.8576% higher than the calculation of MiB/s.

I prefer MiB, so I'm not gonna keep it in the `-v` output, but I'm also
gonna include MB to avoid misunderstandings.