`-runs N` specifies the number of times to run the test
This is especially useful to smooth out disk volatility and express
performance in terms of a range (e.g. 70 MB/s with a standard deviation of
20 MB/s).

Also, test setup for me is expensive (`bosh deploy`, `bosh
deld`), which adds ten minutes to every test run. So it's helpful to get
as many runs in as I can once I've deployed my test VM.

Also I refactored a single-argument `path.Join()`, which is as useful as
a box of rocks. Now it has two arguments & allows me to redact a
subsequent `path.Join()`.