Version is maintained in package `bench`
Previously the version (e.g. 1.0.4) was maintained in the main()
wrapper, but it belonged in the package because that's where all the
code is. Plus, the package should be independent of main().

This is laying the groundwork for the JSON output, and Version needs to
be one of the fields.