⏱  JSON output is in seconds, not nanoseconds
Nanoseconds was a poor choice for output, because people measure in
seconds, not nanoseconds (although nanoseconds was an easy choice
because that's the default output of time.Duration).

Also I had the wrong default for the duration of the IOPS test (it's 15
seconds, not 5).

Slight refactoring: `IOPS{Duration,Operations}` →
`IOPS{Duration,Operations}`.  IOPS is a rate ("per second"), so a
duration seems somewhat inappropriate.

`start_time` was added to the JSON output, both when the benchmark as a
whole was started as well as for each individual run. I like to see when
a given benchmark was run.