Buffer Cache is cleared every three seconds
This helps generate more accurate IOPS & read benchmarks when `-size` is
much less than the system RAM, or `-iops-duration` is long.

For example, on my MacBook Pro with `-size=0.5`, my IOPS result is 495k.
If I clear the buffer cache before the IOPS test, it's 345k.  If I clear
the buffer cache before the IOPS test _and_ every 3 seconds, it's 45k.

Admittedly 3 seconds is arbitrary, and the real number depends on such
factors as how big your test file size is and how fast your disk is,
etc. For a pathological case, assume a test file size of 1 disk block —
the value will be cached after the first operation!