⏲   `-seconds` flag sets duration of benchmark
I'd like to be able to set the duration of a benchmark. Specifically,
I'd like to be able to set the benchmark to run for an entire day in
order to exhaust AWS's [I/O
Credits](https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/EBSVolumeTypes.html#IOcredit)
and determine what performance can be expected from gp2 storage.

`-runs` is not suitable for this task — for fast storage it finishes too
quickly, for slow storage, too slowly.

I've also updated the README sample output to include the new fields.